+++ sdk/formrecognizer/azure-ai-formrecognizer/src/test/java/com/azure/ai/formrecognizer/DocumentAnalysisAsyncClientTest.java
@@ -10,6 +10,7 @@ import com.azure.ai.formrecognizer.models.AnalyzeResult;
 import com.azure.ai.formrecognizer.models.DocumentOperationResult;
 import com.azure.core.exception.HttpResponseException;
 import com.azure.core.http.HttpClient;
+import com.azure.core.models.ResponseError;
 import com.azure.core.util.polling.SyncPoller;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
@@ -189,9 +190,8 @@ public class DocumentAnalysisAsyncClientTest extends DocumentAnalysisClientTestB
                 () -> client.beginAnalyzeDocument("prebuilt-receipt", Utility.toFluxByteBuffer(data), dataLength)
                     .setPollInterval(durationTestMode)
                     .getSyncPoller().getFinalResult());
-            FormRecognizerError errorInformation =
-                (FormRecognizerError) httpResponseException.getValue();
-            Assertions.assertEquals("InvalidRequest", errorInformation.getCode());
+            ResponseError responseError = (ResponseError) httpResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 
@@ -248,8 +248,8 @@ public class DocumentAnalysisAsyncClientTest extends DocumentAnalysisClientTestB
                     .getSyncPoller()
                     .getFinalResult());
 
-            FormRecognizerError errorInformation = (FormRecognizerError) httpResponseException.getValue();
-            Assertions.assertEquals("InvalidRequest", errorInformation.getCode());
+            ResponseError responseError = (ResponseError) httpResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 
@@ -389,8 +389,8 @@ public class DocumentAnalysisAsyncClientTest extends DocumentAnalysisClientTestB
                 () -> client.beginAnalyzeDocument("prebuilt-layout", Utility.toFluxByteBuffer(data), dataLength)
                     .setPollInterval(durationTestMode)
                     .getSyncPoller().getFinalResult());
-            FormRecognizerError errorInformation = (FormRecognizerError) httpResponseException.getValue();
-            Assertions.assertEquals("InvalidRequest", errorInformation.getCode());
+            ResponseError responseError = (ResponseError) httpResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 
@@ -522,9 +522,8 @@ public class DocumentAnalysisAsyncClientTest extends DocumentAnalysisClientTestB
             HttpResponseException httpResponseException = assertThrows(HttpResponseException.class,
                 () -> client.beginAnalyzeDocumentFromUrl("prebuilt-layout", invalidSourceUrl)
                     .setPollInterval(durationTestMode).getSyncPoller().getFinalResult());
-            FormRecognizerError errorInformation =
-                (FormRecognizerError) httpResponseException.getValue();
-            Assertions.assertEquals("InvalidRequest", errorInformation.getCode());
+            ResponseError responseError = (ResponseError) httpResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 
@@ -686,9 +685,8 @@ public class DocumentAnalysisAsyncClientTest extends DocumentAnalysisClientTestB
                     .getSyncPoller()
                     .getFinalResult());
 
-            FormRecognizerError errorInformation =
-                (FormRecognizerError) httpResponseException.getValue();
-            Assertions.assertEquals("InvalidRequest", errorInformation.getCode());
+            ResponseError responseError = (ResponseError) httpResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 
@@ -766,9 +764,8 @@ public class DocumentAnalysisAsyncClientTest extends DocumentAnalysisClientTestB
                     .getSyncPoller()
                     .getFinalResult());
 
-            FormRecognizerError errorInformation =
-                (FormRecognizerError) httpResponseException.getValue();
-            Assertions.assertEquals("InvalidRequest", errorInformation.getCode());
+            ResponseError responseError = (ResponseError) httpResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 
@@ -934,9 +931,8 @@ public class DocumentAnalysisAsyncClientTest extends DocumentAnalysisClientTestB
                     .setPollInterval(durationTestMode)
                     .getSyncPoller()
                     .getFinalResult());
-            FormRecognizerError errorInformation =
-                (FormRecognizerError) httpResponseException.getValue();
-            Assertions.assertEquals("InvalidRequest", errorInformation.getCode());
+            ResponseError responseError = (ResponseError) httpResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 
@@ -1144,8 +1140,8 @@ public class DocumentAnalysisAsyncClientTest extends DocumentAnalysisClientTestB
                     .getSyncPoller()
                     .getFinalResult());
 
-            FormRecognizerError errorInformation = (FormRecognizerError) httpResponseException.getValue();
-            Assertions.assertEquals("InvalidRequest", errorInformation.getCode());
+            ResponseError responseError = (ResponseError) httpResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 
@@ -1182,9 +1178,8 @@ public class DocumentAnalysisAsyncClientTest extends DocumentAnalysisClientTestB
                 = assertThrows(HttpResponseException.class,
                     () -> client.beginAnalyzeDocumentFromUrl("prebuilt-idDocument", invalidSourceUrl)
                         .setPollInterval(durationTestMode).getSyncPoller().getFinalResult());
-            FormRecognizerError errorInformation =
-                (FormRecognizerError) httpResponseException.getValue();
-            Assertions.assertEquals("InvalidRequest", errorInformation.getCode());
+            ResponseError responseError = (ResponseError) httpResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 }