+++ sdk/formrecognizer/azure-ai-formrecognizer/src/test/java/com/azure/ai/formrecognizer/DocumentAnalysisClientTest.java
@@ -5,7 +5,7 @@ package com.azure.ai.formrecognizer;
 
 import com.azure.ai.formrecognizer.administration.DocumentModelAdministrationClient;
 import com.azure.ai.formrecognizer.administration.models.DocumentModel;
-import com.azure.ai.formrecognizer.administration.models.FormRecognizerError;
+import com.azure.ai.formrecognizer.administration.models.DocumentModelOperationError;
 import com.azure.ai.formrecognizer.models.AnalyzeDocumentOptions;
 import com.azure.ai.formrecognizer.models.AnalyzeResult;
 import com.azure.ai.formrecognizer.models.AnalyzedDocument;
@@ -13,6 +13,7 @@ import com.azure.ai.formrecognizer.models.DocumentField;
 import com.azure.ai.formrecognizer.models.DocumentOperationResult;
 import com.azure.core.exception.HttpResponseException;
 import com.azure.core.http.HttpClient;
+import com.azure.core.models.ResponseError;
 import com.azure.core.util.Context;
 import com.azure.core.util.polling.SyncPoller;
 import org.junit.jupiter.api.AfterAll;
@@ -191,9 +192,9 @@ public class DocumentAnalysisClientTest extends DocumentAnalysisClientTestBase {
                     () -> client.beginAnalyzeDocument("prebuilt-receipt", data, dataLength)
                         .setPollInterval(durationTestMode)
                         .getFinalResult());
-            FormRecognizerError errorInformation
-                = (FormRecognizerError) httpResponseException.getValue();
-            Assertions.assertEquals("InvalidContent", errorInformation.getInnerError().getCode());
+            ResponseError responseError =
+                (ResponseError) httpResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 
@@ -378,9 +379,9 @@ public class DocumentAnalysisClientTest extends DocumentAnalysisClientTestBase {
                     () -> client.beginAnalyzeDocument("prebuilt-layout", data, dataLength)
                         .setPollInterval(durationTestMode)
                         .getFinalResult());
-            FormRecognizerError errorInformation
-                = (FormRecognizerError) errorResponseException.getValue();
-            Assertions.assertEquals("InvalidContent", errorInformation.getInnerError().getCode());
+            ResponseError responseError =
+                (ResponseError) errorResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 
@@ -836,12 +837,12 @@ public class DocumentAnalysisClientTest extends DocumentAnalysisClientTestBase {
                         INVALID_URL)
                     .setPollInterval(durationTestMode)
                     .getFinalResult());
-            final FormRecognizerError errorInformation
-                = (FormRecognizerError) httpResponseException.getValue();
+            final ResponseError responseError =
+                (ResponseError) httpResponseException.getValue();
 
             adminClient.deleteModel(createdModel.getModelId());
 
-            Assertions.assertEquals("InvalidContentSourceFormat", errorInformation.getInnerError().getCode());
+            Assertions.assertEquals("InvalidContentSourceFormat", responseError.getCode());
         });
     }
 
@@ -874,9 +875,8 @@ public class DocumentAnalysisClientTest extends DocumentAnalysisClientTestBase {
             HttpResponseException errorResponseException = Assertions.assertThrows(HttpResponseException.class,
                 () -> client.beginAnalyzeDocumentFromUrl(NON_EXIST_MODEL_ID, fileUrl)
                     .setPollInterval(durationTestMode));
-            FormRecognizerError errorInformation
-                = (FormRecognizerError) errorResponseException.getValue();
-            Assertions.assertEquals("ModelNotFound", errorInformation.getInnerError().getCode());
+            ResponseError responseError = (ResponseError) errorResponseException.getValue();
+            Assertions.assertEquals("ModelNotFound", responseError.getCode());
         }, CONTENT_FORM_JPG);
     }
 
@@ -1034,9 +1034,8 @@ public class DocumentAnalysisClientTest extends DocumentAnalysisClientTestBase {
                 () -> client.beginAnalyzeDocument("prebuilt-businessCard", data, dataLength)
                     .setPollInterval(durationTestMode)
                     .getFinalResult());
-            FormRecognizerError errorInformation
-                = (FormRecognizerError) httpResponseException.getValue();
-            Assertions.assertEquals("InvalidContent", errorInformation.getInnerError().getCode());
+            ResponseError responseError = (ResponseError) httpResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 
@@ -1208,9 +1207,8 @@ public class DocumentAnalysisClientTest extends DocumentAnalysisClientTestBase {
                 () -> client.beginAnalyzeDocument("prebuilt-invoice", data, dataLength)
                     .setPollInterval(durationTestMode)
                     .getFinalResult());
-            FormRecognizerError errorInformation
-                = (FormRecognizerError) httpResponseException.getValue();
-            Assertions.assertEquals("InvalidContent", errorInformation.getInnerError().getCode());
+            ResponseError responseError = (ResponseError) httpResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 
@@ -1417,9 +1415,8 @@ public class DocumentAnalysisClientTest extends DocumentAnalysisClientTestBase {
                         dataLength)
                     .setPollInterval(durationTestMode)
                     .getFinalResult());
-            FormRecognizerError errorInformation
-                = (FormRecognizerError) httpResponseException.getValue();
-            Assertions.assertEquals("InvalidContent", errorInformation.getInnerError().getCode());
+            ResponseError responseError = (ResponseError) httpResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 
@@ -1454,9 +1451,8 @@ public class DocumentAnalysisClientTest extends DocumentAnalysisClientTestBase {
                 () -> client.beginAnalyzeDocumentFromUrl("prebuilt-idDocument", invalidSourceUrl)
                     .setPollInterval(durationTestMode)
                     .getFinalResult());
-            FormRecognizerError errorInformation
-                = (FormRecognizerError) errorResponseException.getValue();
-            Assertions.assertEquals("InvalidContent", errorInformation.getInnerError().getCode());
+            ResponseError responseError = (ResponseError) errorResponseException.getValue();
+            Assertions.assertEquals("InvalidContent", responseError.getCode());
         });
     }
 }